Remove cache system
I wanted to cache the request to Workflow Run usages to avoid rate limit
on GitHub. This API endpoint requires a Workflow Run ID to retrieve each
usage of Workflow Run, so we usually reach the rate limit when there are
a lot of Workflow Runs.

Thus, I followed this guideline and used conditional requests to avoid
consumption of rate limit on GitHub.

https://docs.github.com/en/rest/using-the-rest-api/best-practices-for-using-the-rest-api?apiVersion=2022-11-28#use-conditional-requests-if-appropriate

However, this didn't seem to work for token assigned to GitHub Workflow
while it worked as expected with a Personal Access Token.
Maybe, this is a constraint by GitHub, so I gave up using conditional
requests for tokens available on GitHub Workflows.