Do not use Usage and calculate duration from createdAt/updatedAt again
`Usage` requires `getWorkflowRunUsage()`, and it requires an API request
for each GitHub Workflow Runs. Multiple API calls for a single GitHub
Workflow Run easily reaches rate limit on GitHub, so it's hard for
repositories that have many workflow runs to use this GitHub Action.

So, I decided not to use `Usage`. Instead, I decided to calculate
`Duration` from `createdAt` and `updatedAt`. Of course, this is not
correct, but in most cases, it's enough to show the tendency of the
GitHub Workflows.

Maybe, I will remove `getWorkflowRunUsage()` and `Usage`, but in this
commit, I remain them. The decision to remove them should be done later.