Use run_started_at to calculate duration
`run_started_at` is updated when `run_attempt` is more than `1`.
This means when a workflow run is re-run, we can calculate the duration
of the latest attempt for the workflow run. It must be more accurate
than using `created_at`.

Note `GitHubWorkflowRun.date` still returns a `Date` instance based on
`created_at` because `date` should be consistent. If `date` will change
because of re-run, a chart order will change. This is odd.